+++ tool/gec/backend/c/runtime/eif_config.h
@@ -1,0 +1,17 @@
+/*
+	description:
+
+		"Part of ISE Eiffel runtime. Needed to compile the C code of the EiffelNet library."
+
+	system: "Gobo Eiffel Compiler"
+	copyright: "Copyright (c) 2024, Eric Bezault and others"
+	license: "MIT License"
+*/
+
+#ifndef EIF_CONFIG_H
+#define EIF_CONFIG_H
+#if defined(_MSC_VER) && (_MSC_VER >= 1020)
+#pragma once
+#endif
+
+#endif